requirements.txt
python-keyczar
